requirements.txt
numpy==2.3.0
sentencepiece==0.2.0
torch==2.7.1
transformers==4.52.4
pillow==11.2.1